Support for implementing Registration UC
THe significant functional change was to allow the Registration page
to be pushed from the Home page.  This allowed the Registration process
to complete without having to know where it was being pushed from -- pop
up when the process was complete.

Also contains back-versioning of Angular from 5.0 back to 4.1 until the
Page Title thing gets resolved.  (This broke all my BDD tests that
depended on seeing the title of a page -- which was all of them.)