PLAY-44 Place an Edge on the map
- Adds EdgeService to provide the Rolling Map with a track to place.
- Amends the HTML template to place the map.
- Changes the CSS to provide full height to the container.
- Removes padding from the page so map can fill entire screen.

This hardcodes an initial position for the map instead of centering on content. Once the patterns for bringing across content settle out, this will change to centering on that content.

This does introduce an "edgeLayer", that we'll want to get more sophisticated about. The Subscribe model works well to push those
changes. Will be interesting to see how it plays out in the field.