Exploring how we respond to Game State to present Paths on the map
- Adds lots of logging to provide life-cycle visibility.
- Begins subscriptions to GameState events at the time user decides to begin the game.
- GameStateService responsible for providing Subject that subscribers can listen to for these changes. Rolling Page is current subscriber.
- Passes Game State into rolling-map component via
  - Subscription in Service,
  - @Input in component,
  - ngOnChanges() within component controller

Learning that our navigation choices are limited. When "Setting Root", a new instance of the page is created. We only want a single instance of our pages. This commit is limited in what it can do for that reason because each transition to the Rolling/Map page adds a new copy to the subscriber list.